Rework of wingclassification in another script file
